psycopg2_sample/requirements.txt
alpaca_trade_api
psycopg2-binary
requests 
asyncpg 
aiohttp
tqdm